# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AndersSpotify API",
        "description": "Documentación de la API para el proyecto de reproducción musical tipo Spotify",
        "contact": {
            "email": "fernandaurcelay5@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/music/{id}/favorite": {
            "post": {
                "tags": [
                    "Música"
                ],
                "summary": "Agregar o quitar una canción de favoritos (toggle)",
                "operationId": "4d7ec90bc79912319a3d7fbc5b0e6654",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la canción a marcar o desmarcar como favorita",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado de la acción",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Canción eliminada de favoritos"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "music_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Canción no encontrada o no autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Música"
                ],
                "summary": "Obtener canciones favoritas del usuario autenticado",
                "operationId": "862c6428f4eb267ee67a0dabeaf5a018",
                "responses": {
                    "200": {
                        "description": "Lista de canciones favoritas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 7
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Uptown Funk"
                                            },
                                            "file_url": {
                                                "type": "string",
                                                "example": "https://..."
                                            },
                                            "cover_image": {
                                                "type": "string",
                                                "example": "https://..."
                                            },
                                            "duration": {
                                                "type": "string",
                                                "example": "00:04:31"
                                            },
                                            "play_count": {
                                                "type": "integer",
                                                "example": 700
                                            },
                                            "likes": {
                                                "type": "integer",
                                                "example": 300
                                            },
                                            "artist": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Bruno Mars"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "album": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "24K Magic"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "category": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Funk"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Usa el token generado tras el login. Formato: Bearer (token)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}